Client side setup

+++ server/controllers/recipes.js
@@ -1,14 +1,8 @@
 "use strict"
-import multer from 'multer';
 import path from 'path';
 
 import validate from '../middleware/validate';
 import db from '../models';
-
-const upload = multer({
-	dest : '../public'
-})
-
 
 class Recipes {
 
@@ -19,6 +13,9 @@ class Recipes {
   
   getAllRecipes(req, res) {
     // Get recipes based or query strings to return the most voted recipes
+    let limit = 6;   // number of records per page
+    let offset = 0;
+
     if (req.query.sort && req.query.order) {
       let order = req.query.order;
       if (req.query.sort === 'upvotes') {
@@ -29,6 +26,7 @@ class Recipes {
         }
         db.Recipes.findAll({
             limit: 10,
+            offset: offset,
             order: [
               ['upvotes', order]
             ]
@@ -47,7 +45,8 @@ class Recipes {
           order = 'ASC';
         }
         db.Recipes.findAll({
-            limit: 10,
+            limit : limit,
+            offset: offset,
             order: [
               ['upvotes', order]
             ]
@@ -62,13 +61,9 @@ class Recipes {
     }
 
     //get all recipes
-
-    let limit = 6;   // number of records per page
-    let offset = 0;
     db.Recipes.findAndCountAll()
     .then((data) => {
-      let page = req.body.page ? req.body.page : 2;      // page number
-      console.log(req.body);
+      let page = req.body.page ? req.body.page : 1;      // page number
       let pages = Math.ceil(data.count / limit);
       offset = limit * (page - 1);
 
@@ -81,12 +76,10 @@ class Recipes {
       })
       .then((recipes) => {
         if(recipes){
-          // console.log(recipes)
           res.status(200).json({status: 200, data: recipes, pages : pages });
         }else{
-          res.status(401).json({status: 401, message: "No recipe found"});
+          res.status(401).json({status: 400, message: "No recipe found"});
         }
-
       })
       .catch(error => {
         res.status(500).json(error);
@@ -100,49 +93,46 @@ class Recipes {
 	*
 	*/
   getSingleRecipe(req, res) {
-    if(validate.validateId(req.params.id)){ 
-      db.Recipes.findById(req.params.id)
+    const id = req.params.id;
+    if(validate.validateId(id)){ 
+      db.Recipes.findById(id)
         .then((recipe) => {
           if(recipe){
             res.status(200).json({status: 200, data: recipe});
           }else{
-            res.status(401).json({status: 401, message: "Recipe with id '" + req.params.id + "' not found"});
+            res.status(400).json({status: 400, message: "Recipe with id '" + id + "' not found"});
           }          
         })
         .catch(error => {
           res.status(500).json({status: 500, message: error.message});
         })
     }else{
-      res.status(400).json({message: `Id "${req.params.id}" is not a valid integer` });
+      res.status(400).json({message: `Id "${id}" is not a valid integer` });
     }
   }
 
-    /*
+  /*
 	* getSingleRecipes function with @params id, and a return type of array
 	*
 	*/
   getMyRecipes(req, res) {
-    
-    console.log(req.params.id);
-    console.log(validate.validateId(req.params.id));
-
-    if(validate.validateId(req.params.id)){ 
+    const id = req.params.id;
+    if(validate.validateId(id)){ 
       let limit = 6;   // number of records per page
       let offset = 0;
       db.Recipes.findAndCountAll({ 
         where : { 
-            userId : req.params.id
+            userId : id
           }
         })
       .then((data) => {
         let page = req.body.page ? req.body.page : 1;      // page number
-        console.log(req.body);
         let pages = Math.ceil(data.count / limit);
         offset = limit * (page - 1);
 
         db.Recipes.findAll({ 
           where : { 
-            userId : req.params.id
+            userId : id
           },
           limit: limit,
           offset: offset,
@@ -154,11 +144,8 @@ class Recipes {
             if(recipes){
               res.status(200).json({status: 200, data: recipes});
             }else{
-              res.status(401).json({status: 401, message: "User with id '" + req.params.userid + "' not found"});
+              res.status(400).json({status: 400, message: "User with id '" + rid + "' not found"});
             }          
-          })
-          .catch(error => {
-            res.status(500).json({status: 500, message: error.message});
           })
         })
         .catch(err =>{
@@ -174,24 +161,21 @@ class Recipes {
 	*
 	*/
   addRecipe(req, res) {
-    
     validate.validateAddRecipes(req, res);
     var errors = req.validationErrors();
     if (errors) {
-      res.status(400).json({ status : 400, message : errors });
-      return;
+      return res.status(400).json({ status : 400, message : errors });
     } else {
       db.Recipes.create(req.body)
         .then((result) => {
           if (result) {
             return res.status(201).json({status: 201, message: "Recipe added"});
           } else {
-            return res.status(400).json({status: 400, message: "no result returned"});
+            return res.status(400).json({status: 400, message: "Error Adding recipe, try again later"});
           }
         })
         .catch(error => {
-          console.log(error);
-          // return res.status(500).json({status: 500, message: error.message});
+          return res.status(500).json({status: 500, message: error.message});
         })
     }
   }
@@ -202,8 +186,14 @@ class Recipes {
 	*/
   deleteRecipe(req, res) {
     const id = req.params.id;
-    if(validate.validateId(id)){ 
-      db.Recipes.findById(id)
+    const userId = req.body.userId;
+    if(validate.validateId(id) && validate.validateId(userId)){ 
+      db.Recipes.findOne({
+        where : {
+          id : id,
+          userId : userId
+        }
+      })
       .then(result => {
         if(result){
           db.Recipes.destroy({
@@ -220,23 +210,27 @@ class Recipes {
             res.status(400).json({message: err.message});
           });
         } else{
-          res.status(400).json({message: "Recipe with id "+ req.params.id +" not found"});
+          res.status(400).json({message: 'User with id ' + userId + ' has no recipe with id '+ id });
         }  
       }); 
     }else{
-      res.status(400).json({message: `Id "${req.params.id}" is not a valid integer` });
+      res.status(400).json({message: `User ID and recipe ID must be a valid integer` });
     }
   }
 
   updateRecipe(req, res) {
-    validate.validateAddRecipes(req, res);
+    const id = req.params.id;
+    const userId = req.body.userId;   
     var errors = req.validationErrors();
     if (errors) {
-      res.status(400).json({ status : 400, message : errors });
-      return;
+      return res.status(400).json({ status : 400, message : errors });
     } else {
-      if(validate.validateId(req.params.id)){ 
-        db.Recipes.findById(req.params.id)
+      if(validate.validateId(id) && validate.validateId(userId)){ 
+        db.Recipes.findOne({
+          where : {
+            id, userId 
+          }
+        })
         .then(result => {
           if(result){
             db.Recipes.update(req.body, {
@@ -254,49 +248,67 @@ class Recipes {
             });
           }
           else{
-            res.status(400).json({message: "Recipe with id "+ req.params.id +" not found"});
+            res.status(400).json({message: "Recipe with id "+ id +" not found"});
           }
         })
       }else{
-        res.status(400).json({message: `Id "${req.params.id}" is not a valid integer` });
+        res.status(400).json({ status : 400, message : 'user ID and recipe ID must be a valid integer' });
       }
-    }
+      }
     
   }
 
   reviewRecipe(req, res) {
-    db.Reviews.create(req.body)
-      .then((result) => {
-        res.status(200).json({status: 200, data: result});
-      })
-      .catch(error => {
-        res.status(500).json({ status : 500, message : error.message });
-      })
+    const { recipeId, userId, comment } = req.body;
+    validate.validateReview(req, res);
+    var errors = req.validationErrors();
+    if (errors) {
+      return res.status(400).json({ status : 400, message : errors });
+    }
+    db.Reviews.create({
+      recipeId, userId, comment
+    })
+    .then((result) => {
+      res.status(200).json({ status: 200, data: result});
+    })
+    .catch(error => {
+      res.status(500).json({ status : 500, message : error.message });
+    })
   }
 
   getReviews(req, res) {
+    const id = req.params.id;
+    if(!validate.validateId(id)){
+      req.status(400).json({ status : 400, message : "Id is not a valid integer"});
+    }
     db.Reviews.findAll({
-        where: {
-          recipeId: req.params.id
-        }
-      })
-      .then((result) => {
-        if (result) {
-          res.status(200).json({status: 200, data: result});
-        }
-
-      })
-      .catch(error => {
-        res.status(500).json({ status : 500, message : error.message });
-      })
+      where: {
+        ecipeId: id
+      }
+    })
+    .then((result) => {
+      if (result) {
+        res.status(200).json({ status: 200, data: result});
+      }
+      else{
+        res.status(400).json({ status: 400, message: 'No review found for this recipe'});
+      }
+    })
+    .catch(error => {
+      res.status(500).json({ status : 500, message : error.message });
+    })
   }
 
   addUpvote(req, res) {
-
     const userId = req.body.userId;
-    const recipeId = req.params.id;
+    const id = req.params.id;
     const vote = req.body.vote;
-
+    if(validate.validateId(userId) && validate.validateId(id)) {
+      return res.status(400).json({ status : 400, message : 'user ID or Recipe ID is invalid' });
+    }
+    if(validate.validateId(userId) && validate.validateId(id)) {
+      return res.status(400).json({ status : 400, message : 'user ID or Recipe ID is invalid' });
+    }
     db.Votes.findOne({
       where: {
         recipeId: recipeId,
@@ -304,14 +316,14 @@ class Recipes {
       }
     }).then(res => {
       if(res){
-       console.log("user already voted")
-       return;
+        return res.status(400).json({ status : 400, message : 'Already Votted' });
       }
       db.Recipes.findOne({
         where: {
           recipeId: recipeId
         }
-      }).then(recipe =>{
+      })
+      .then(recipe =>{
         if(recipe){
           const newVote = recipe.upvote + 1;
           db.Recipes.update({ upvote : newVote }, {
@@ -322,9 +334,9 @@ class Recipes {
           .then(res => {
             db.Votes.create({
               recipeId: recipeId,
-              userId : userId
+              userId : userId,
+              voteType : vote
             }).then( res => {
-              
             }) 
           })
         }
@@ -361,43 +373,48 @@ class Recipes {
   }
 
   getFavourites(req, res) {
-		db.Favourites.findAll({ attributes : ['recipeId']}, { where : { userId : req.params.id}})
-		.then((result) => {	
-      let ids = [];
-      for(let i = 0; i < result.length; i++){
-        ids.push(result[i].recipeId);
-      }
+    const userId = req.params.id;
+    if(!validate.validateId(id)){
+      req.status(400).json({ status : 400, message : "Id is not a valid integer"});
+    }
+		db.Favourites.findAll({ attributes : ['recipeId']}, { where : { userId  }})
+		.then((favourite) => {
+      if(favourite){
+        let ids = [];
+        result.map(favourite => {
+          ids.push(favourite.recipeId);
+        });
+      }	
       console.log(ids);
-     
       db.Recipes.findAll({ 
         where : { id : [...ids]}
       })
       .then(favourites => {
-        res.status(200).json({ status: 200, data : favourites });
+        if (favourites ) {
+         res.status(200).json({ status: 200, data : favourites });
+        }else{
+          res.status(401).json({ status: 401, message : "No recipe found!!" });
+        } 
       })
-    //   if (favourites ) {
-    //      res.status(200).json({ status: 200, data : favourites });
-		// 	}else{
-    //     res.status(401).json({ status: 401, message : "No recipe found!!" });
-    //   } 
-		// })
-		// .catch(err => {
-		// 	res.status(500).json({ status: 500, message : err.message });
+		})
+		.catch(err => {
+			res.status(500).json({ status: 500, message : err.message });
 		});
 	}
 
 	addFavourites(req, res) {
     const userId = req.params.id;
     const recipeId = req.body.recipeId;
-    console.log(userId, recipeId);
+    if(validate.validateId(userId) && validate.validateId(recipeId)){
+      res.status(400).json({ status : 400, message: "User Id and Recipe ID must be a valid not integer"});
+    }
     const data = {
-      userId : userId,
-      recipeId : recipeId
+      userId, recipeId
     }
 		db.Favourites.create(data)
-		.then((favourites) => {			
-      if (favourites ) {
-         res.status(201).json({ status: 201, data : favourites });
+		.then((favourite) => {			
+      if (favourite) {
+         res.status(201).json({ status: 201, data : favourite });
 			}  
 		})
 		.catch(err => {
@@ -406,9 +423,10 @@ class Recipes {
   }
   
   removeFavourites(req, res) {
-    console.log(req.body.recipeId)
-    if(validate.validateId(req.body.recipeId)){ 
-      db.Favourites.findOne({ where : { userId : req.params.id, recipeId : req.body.recipeId}})
+    const recipeId = req.body.recipeId;
+    const userId = req.params.id;
+    if(validate.validateId(recipeId)){ 
+      db.Favourites.findOne({ where : { userId, recipeId }})
       .then(result => {
         if(result){
           db.Favourites.destroy({
@@ -420,14 +438,14 @@ class Recipes {
             if (result) {
               res.status(200).json({message: "Favourite recipe removed"});
             }else{
-              res.status(400).json({ status : 400, message: "Recipe with id "+ req.params.id +" not found"});
+              res.status(400).json({ status : 400, message: "Recipe with id "+ recipeId +" not found"});
             }
           })
           .catch(err => {
             res.status(400).json({ status : 400, message: "cannot find recipes"});
           });
         } else{
-          res.status(400).json({ status : 400, message: "Recipe with id "+ req.params.id +" not found"});
+          res.status(400).json({ status : 400, message: "Recipe with id "+ recipeId +" not found"});
         }  
       }); 
     }else{
@@ -440,15 +458,15 @@ class Recipes {
         where: {
           id: req.params.id
         }
-      })
-      .then((items) => {
-        if (items) {
-          res.status(200).json({status: 200, data: items});
-        }
-      })
-      .catch(err => {
-        res.status(400).json({message: err.message});
-      });
+    })
+    .then((items) => {
+      if (items) {
+        res.status(200).json({status: 200, data: items});
+      }
+    })
+    .catch(err => {
+      res.status(400).json({message: err.message});
+    });
   }
 }
